refactor(open-mpm): split compress/ (tool_output + mod.rs pipeline) (#366)
