--- crates/open-mpm/src/compress/tool_output.rs
@@ -1,1002 +1,0 @@
-//! Per-tool output compression filters.
-//!
-//! Why: Verbose tool outputs (cargo test, git diff, git log, file reads) waste
-//! tokens when re-injected into LLM conversation history. Stripping noise
-//! preserves signal while shrinking context.
-//! What: `compress_tool_output(name, output)` dispatches to a filter based on
-//! the tool name, returning a possibly-shorter string. Each filter is a pure
-//! `fn` for unit testability.
-//! Test: See module-level `tests` — covers each filter and the dispatch table.
-
-/// Minimum byte length below which compression is skipped.
-///
-/// Why: Tiny outputs (status codes, short results) are already cheaper than
-/// the cognitive cost of compression artifacts. RTK uses 80 bytes.
-const SIZE_GATE_BYTES: usize = 80;
-
-/// Compress a tool's textual output based on its name.
-///
-/// Why: Centralizes the per-tool filter dispatch so callers don't have to
-/// know which filter applies to which tool.
-/// What: Routes by substring match in `tool_name`. Applies a size gate and a
-/// structured-format passthrough before any filter runs. Unknown tools pass
-/// through unchanged. Always infallible.
-/// Test: `compress_tool_output_dispatch_test` plus per-filter tests.
-pub fn compress_tool_output(tool_name: &str, output: &str) -> String {
-    // Size gate: very small outputs aren't worth touching.
-    if output.len() < SIZE_GATE_BYTES {
-        return output.to_string();
-    }
-    // Structured formats (JSON/YAML/TOML/CSV) must pass through unchanged so
-    // we don't corrupt machine-parseable payloads.
-    if is_structured_format(output) {
-        return output.to_string();
-    }
-    let n = tool_name.to_ascii_lowercase();
-    if n.contains("test") || n.contains("cargo") {
-        // Note: "cargo check"/"cargo clippy" go to the cargo_check filter
-        // which strips Compiling/Finished lines; "cargo test" goes here.
-        if n.contains("check") || n.contains("clippy") {
-            return filter_cargo_check(output);
-        }
-        return filter_test_runner(output);
-    }
-    if n.contains("diff") {
-        return filter_git_diff(output);
-    }
-    if n.contains("log") {
-        let line_count = output.lines().count();
-        if line_count > 30 {
-            return filter_git_log(output);
-        }
-        return output.to_string();
-    }
-    if n.contains("read") || n.contains("cat") {
-        let line_count = output.lines().count();
-        if line_count > 200 {
-            return filter_file_read(output);
-        }
-        return output.to_string();
-    }
-    if n.contains("check") || n.contains("clippy") {
-        return filter_cargo_check(output);
-    }
-    output.to_string()
-}
-
-/// Strip passing test lines from `cargo test` output.
-///
-/// Why: Hundreds of `test foo ... ok` lines drown out the few failures the
-/// model needs to see.
-/// What: Drops lines matching `test <name> ... ok`. Keeps `FAILED`, `error`,
-/// `warning`, and the final `test result:` summary. Returns summary-only
-/// when nothing else remains.
-/// Test: `test_runner_strips_passing_tests`, `test_runner_keeps_summary_line`,
-/// `test_runner_no_failures_returns_summary_only`.
-pub fn filter_test_runner(output: &str) -> String {
-    let mut kept: Vec<&str> = Vec::new();
-    let mut summary: Option<&str> = None;
-    for line in output.lines() {
-        let trimmed = line.trim_start();
-        if trimmed.starts_with("test result:") {
-            summary = Some(line);
-            continue;
-        }
-        if is_passing_test_line(trimmed) {
-            continue;
-        }
-        let lower = trimmed.to_ascii_lowercase();
-        if trimmed.contains("FAILED")
-            || lower.contains("error")
-            || lower.contains("warning")
-            || trimmed.starts_with("---- ")
-            || trimmed.starts_with("failures:")
-        {
-            kept.push(line);
-        }
-    }
-
-    if kept.is_empty() {
-        return summary.map(|s| s.to_string()).unwrap_or_default();
-    }
-    let mut out = kept.join("\n");
-    if let Some(s) = summary {
-        out.push('\n');
-        out.push_str(s);
-    }
-    out
-}
-
-fn is_passing_test_line(s: &str) -> bool {
-    // Match `test <something> ... ok` (cargo's per-test line)
-    if !s.starts_with("test ") {
-        return false;
-    }
-    s.ends_with(" ... ok") || s.ends_with(" ... ignored")
-}
-
-/// Collapse runs of context lines in a unified diff.
-///
-/// Why: Diff context (lines starting with a space) is rarely needed by the
-/// model; the +/- lines carry the real signal.
-/// What: Replaces every run of context lines with a single
-/// `@@ ... @@ [+N added, -N removed]` summary header reflecting counts of
-/// the surrounding hunk. Preserves `---`, `+++`, `@@`, `+`, `-` lines.
-/// Test: `filter_git_diff_strips_context_lines`,
-/// `filter_git_diff_preserves_adds_and_removes`,
-/// `filter_git_diff_passthrough_no_context`.
-pub fn filter_git_diff(output: &str) -> String {
-    // First pass: count adds/removes per hunk so we can annotate replacement headers.
-    let lines: Vec<&str> = output.lines().collect();
-    let mut hunk_stats: Vec<(usize, usize)> = Vec::new(); // (added, removed) per hunk
-    let mut cur_add = 0usize;
-    let mut cur_rem = 0usize;
-    let mut in_hunk = false;
-    for line in &lines {
-        if line.starts_with("@@") {
-            if in_hunk {
-                hunk_stats.push((cur_add, cur_rem));
-            }
-            cur_add = 0;
-            cur_rem = 0;
-            in_hunk = true;
-            continue;
-        }
-        if !in_hunk {
-            continue;
-        }
-        if line.starts_with("+++") || line.starts_with("---") {
-            continue;
-        }
-        if line.starts_with('+') {
-            cur_add += 1;
-        } else if line.starts_with('-') {
-            cur_rem += 1;
-        }
-    }
-    if in_hunk {
-        hunk_stats.push((cur_add, cur_rem));
-    }
-
-    // Second pass: emit, collapsing context runs.
-    let mut out: Vec<String> = Vec::new();
-    let hunk_idx: usize = 0;
-    let mut in_context_run = false;
-    let mut had_any_context = false;
-    for line in &lines {
-        if line.starts_with("@@") || line.starts_with("---") || line.starts_with("+++") {
-            in_context_run = false;
-            out.push((*line).to_string());
-            continue;
-        }
-        if line.starts_with('+') || line.starts_with('-') {
-            in_context_run = false;
-            out.push((*line).to_string());
-            continue;
-        }
-        // Treat any other line (including " ..." context, blank) as context.
-        had_any_context = true;
-        if !in_context_run {
-            in_context_run = true;
-            // Emit a collapsed-context marker referencing the current hunk's totals.
-            let (a, r) = hunk_stats
-                .get(hunk_idx.saturating_sub(0))
-                .copied()
-                .unwrap_or((0, 0));
-            // Hunk index advances when we encounter an `@@` header; for context
-            // lines belonging to the current hunk we reuse the current totals.
-            let _ = hunk_idx; // explicit no-op to avoid unused warnings if logic shifts.
-            out.push(format!("@@ ... @@ [+{a} added, -{r} removed]"));
-        }
-        // Drop the context line itself.
-        let _ = line;
-    }
-
-    if !had_any_context {
-        // Passthrough: nothing to compress.
-        return output.to_string();
-    }
-    out.join("\n")
-}
-
-/// Strip author/date/body lines from `git log` output, keeping commit headers.
-///
-/// Why: Long logs are mostly metadata; the SHA + subject is enough for
-/// most LLM reasoning.
-/// What: Keeps lines matching `commit <7+ hex chars>` (and the very next
-/// non-blank non-metadata line as the subject). Drops Author:, Date:,
-/// Merge:, and indented body lines.
-/// Test: `filter_git_log_strips_author_date`, `filter_git_log_passthrough_short`.
-pub fn filter_git_log(output: &str) -> String {
-    let mut out: Vec<String> = Vec::new();
-    let mut expect_subject = false;
-    for line in output.lines() {
-        if is_commit_header(line) {
-            out.push(line.to_string());
-            expect_subject = true;
-            continue;
-        }
-        if expect_subject {
-            // Subject is the first indented non-metadata line after author/date.
-            let trimmed = line.trim();
-            if trimmed.is_empty() {
-                continue;
-            }
-            if trimmed.starts_with("Author:")
-                || trimmed.starts_with("Date:")
-                || trimmed.starts_with("Merge:")
-                || trimmed.starts_with("commit ")
-            {
-                continue;
-            }
-            // First real content line — treat as subject.
-            out.push(line.to_string());
-            expect_subject = false;
-        }
-    }
-    out.join("\n")
-}
-
-fn is_commit_header(line: &str) -> bool {
-    // `commit <hash>` where hash is 7+ hex chars.
-    let stripped = match line.strip_prefix("commit ") {
-        Some(s) => s.trim(),
-        None => return false,
-    };
-    let hex_len = stripped
-        .chars()
-        .take_while(|c| c.is_ascii_hexdigit())
-        .count();
-    hex_len >= 7
-}
-
-/// Strip blank and comment-only lines from large file reads.
-///
-/// Why: When the model loads a 500-line file, blanks and bare comments are
-/// often non-essential.
-/// What: Drops lines that are blank, or trim-start with `//` or `#`. If the
-/// result is < 20 lines (over-filtered, perhaps an all-comment file), returns
-/// the original to avoid hiding too much.
-/// Test: `filter_file_read_strips_blank_comment_lines`,
-/// `filter_file_read_no_over_filter`.
-pub fn filter_file_read(output: &str) -> String {
-    let kept: Vec<&str> = output
-        .lines()
-        .filter(|line| {
-            let t = line.trim_start();
-            if t.is_empty() {
-                return false;
-            }
-            if t.starts_with("//") {
-                return false;
-            }
-            if t.starts_with('#') {
-                return false;
-            }
-            true
-        })
-        .collect();
-    if kept.len() < 20 {
-        return output.to_string();
-    }
-    kept.join("\n")
-}
-
-/// Strip `Compiling` and `Finished` chatter from `cargo check`/`cargo clippy`.
-///
-/// Why: These lines are progress noise; warnings/errors are the signal.
-/// What: Drops lines beginning with `   Compiling ` or `    Finished `.
-/// Keeps everything else verbatim.
-/// Test: `filter_cargo_check_strips_compiling`, `filter_cargo_check_keeps_warnings`.
-pub fn filter_cargo_check(output: &str) -> String {
-    output
-        .lines()
-        .filter(|line| !line.starts_with("   Compiling ") && !line.starts_with("    Finished "))
-        .collect::<Vec<_>>()
-        .join("\n")
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Structured-format detection
-// ─────────────────────────────────────────────────────────────────────────────
-
-/// Detect whether `content` is a structured machine-parseable format.
-///
-/// Why: Running line-based filters over JSON/YAML/TOML/CSV mangles structure
-/// and breaks downstream parsers. Structured payloads must round-trip
-/// byte-for-byte.
-/// What: Heuristic detection — checks for JSON braces/brackets, YAML doc
-/// markers or `key:` lines, TOML `[section]` headers, or consistent CSV
-/// comma counts across multiple lines.
-/// Test: `is_structured_format_*` in test module.
-pub fn is_structured_format(content: &str) -> bool {
-    let trimmed = content.trim_start();
-    if trimmed.is_empty() {
-        return false;
-    }
-
-    // JSON: starts with `{` or `[`.
-    let first = trimmed.as_bytes()[0];
-    if first == b'{' || first == b'[' {
-        return true;
-    }
-
-    // YAML: explicit doc marker.
-    if trimmed.starts_with("---\n") || trimmed == "---" || trimmed.starts_with("---\r") {
-        return true;
-    }
-
-    // TOML: first non-comment, non-blank line is `[section]` or `[[array]]`.
-    if let Some(line) = first_meaningful_line(trimmed)
-        && line.starts_with('[')
-        && (line.ends_with(']') || line.contains("]\n"))
-    {
-        return true;
-    }
-
-    // YAML / key:value heuristic — first meaningful line is `key: value`
-    // (key is alphanumeric/underscore/dash, colon followed by space or EOL).
-    if let Some(line) = first_meaningful_line(trimmed)
-        && looks_like_yaml_kv(line)
-    {
-        return true;
-    }
-
-    // CSV: at least 3 non-empty lines, all with the same nonzero comma count.
-    if looks_like_csv(trimmed) {
-        return true;
-    }
-
-    false
-}
-
-fn first_meaningful_line(s: &str) -> Option<&str> {
-    s.lines()
-        .map(str::trim)
-        .find(|l| !l.is_empty() && !l.starts_with('#'))
-}
-
-fn looks_like_yaml_kv(line: &str) -> bool {
-    let bytes = line.as_bytes();
-    let mut i = 0;
-    while i < bytes.len() {
-        let b = bytes[i];
-        if b == b':' {
-            // Must have key chars before, and either EOL or whitespace after.
-            if i == 0 {
-                return false;
-            }
-            let after_ok = i + 1 == bytes.len() || bytes[i + 1] == b' ' || bytes[i + 1] == b'\t';
-            return after_ok;
-        }
-        if !(b.is_ascii_alphanumeric() || b == b'_' || b == b'-' || b == b'.') {
-            return false;
-        }
-        i += 1;
-    }
-    false
-}
-
-fn looks_like_csv(s: &str) -> bool {
-    let lines: Vec<&str> = s.lines().filter(|l| !l.trim().is_empty()).take(8).collect();
-    if lines.len() < 3 {
-        return false;
-    }
-    let count = lines[0].matches(',').count();
-    if count == 0 {
-        return false;
-    }
-    lines.iter().all(|l| l.matches(',').count() == count)
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// FilterLevel / Language / FilterStrategy (ported from RTK, MIT)
-// ─────────────────────────────────────────────────────────────────────────────
-
-/// Filter aggressiveness level.
-///
-/// Why: Different content tolerates different filtering. Code may want
-/// `Minimal` (preserve comments); large logs may want `Aggressive`.
-/// What: Three levels — None passes through, Minimal removes blank/whitespace,
-/// Aggressive also strips line comments by language.
-/// Test: `filter_strategy_*` tests.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum FilterLevel {
-    None,
-    Minimal,
-    Aggressive,
-}
-
-/// Source language for comment-aware filtering.
-///
-/// Why: Stripping `//` comments from Python or `#` comments from Rust is wrong.
-/// Aggressive filtering needs to know the language to use the right syntax.
-/// What: Enum of supported languages with `from_extension` and
-/// `comment_prefix` / `block_comment` helpers.
-/// Test: `language_from_extension_known`, `language_comment_prefix_rust`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Language {
-    Rust,
-    Python,
-    JavaScript,
-    TypeScript,
-    Go,
-    Shell,
-    Data,
-    Unknown,
-}
-
-impl Language {
-    /// Map a file extension (without leading dot) to a `Language`.
-    pub fn from_extension(ext: &str) -> Self {
-        match ext.trim_start_matches('.').to_ascii_lowercase().as_str() {
-            "rs" => Self::Rust,
-            "py" | "pyi" => Self::Python,
-            "js" | "mjs" | "cjs" | "jsx" => Self::JavaScript,
-            "ts" | "tsx" => Self::TypeScript,
-            "go" => Self::Go,
-            "sh" | "bash" | "zsh" => Self::Shell,
-            "json" | "yaml" | "yml" | "toml" | "csv" => Self::Data,
-            _ => Self::Unknown,
-        }
-    }
-
-    /// Line-comment prefix for this language, if any.
-    pub fn comment_prefix(&self) -> Option<&'static str> {
-        match self {
-            Self::Rust | Self::JavaScript | Self::TypeScript | Self::Go => Some("//"),
-            Self::Python | Self::Shell => Some("#"),
-            Self::Data | Self::Unknown => None,
-        }
-    }
-
-    /// Block-comment open/close delimiters for this language, if any.
-    pub fn block_comment(&self) -> Option<(&'static str, &'static str)> {
-        match self {
-            Self::Rust | Self::JavaScript | Self::TypeScript | Self::Go => Some(("/*", "*/")),
-            Self::Python => Some(("\"\"\"", "\"\"\"")),
-            _ => None,
-        }
-    }
-}
-
-/// Strategy trait for content filtering.
-///
-/// Why: Lets callers swap filtering policy (None / Minimal / Aggressive)
-/// without branching at each call site.
-/// What: One method, `filter(&self, content, lang) -> String`. Implementors
-/// are stateless and `Send + Sync` so they can be cached/shared.
-/// Test: `filter_strategy_no_filter_identity`, `filter_strategy_minimal_drops_blanks`,
-/// `filter_strategy_aggressive_strips_rust_line_comments`.
-pub trait FilterStrategy: Send + Sync {
-    fn filter(&self, content: &str, lang: Language) -> String;
-}
-
-/// Pass-through filter — returns content unchanged.
-pub struct NoFilter;
-
-impl FilterStrategy for NoFilter {
-    fn filter(&self, content: &str, _lang: Language) -> String {
-        content.to_string()
-    }
-}
-
-/// Minimal filter — removes blank lines and trailing whitespace.
-pub struct MinimalFilter;
-
-impl FilterStrategy for MinimalFilter {
-    fn filter(&self, content: &str, _lang: Language) -> String {
-        content
-            .lines()
-            .map(|l| l.trim_end())
-            .filter(|l| !l.is_empty())
-            .collect::<Vec<_>>()
-            .join("\n")
-    }
-}
-
-/// Aggressive filter — minimal + strips line comments by language.
-pub struct AggressiveFilter;
-
-impl FilterStrategy for AggressiveFilter {
-    fn filter(&self, content: &str, lang: Language) -> String {
-        let prefix = lang.comment_prefix();
-        let mut out: Vec<&str> = Vec::new();
-        for line in content.lines() {
-            let trimmed = line.trim_end();
-            if trimmed.is_empty() {
-                continue;
-            }
-            if let Some(p) = prefix {
-                let leading = trimmed.trim_start();
-                if leading.starts_with(p) {
-                    continue;
-                }
-            }
-            out.push(trimmed);
-        }
-        // Collapse runs of equivalent lines? Aggressive just drops; keep simple.
-        out.join("\n")
-    }
-}
-
-/// Get a `FilterStrategy` impl for the requested level.
-///
-/// Why: Lets callers obtain a strategy without naming the concrete type.
-/// What: Boxed trait object, one allocation per call (cheap, infrequent).
-/// Test: `get_filter_returns_expected_type`.
-pub fn get_filter(level: FilterLevel) -> Box<dyn FilterStrategy> {
-    match level {
-        FilterLevel::None => Box::new(NoFilter),
-        FilterLevel::Minimal => Box::new(MinimalFilter),
-        FilterLevel::Aggressive => Box::new(AggressiveFilter),
-    }
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// RTK subprocess fallback
-// ─────────────────────────────────────────────────────────────────────────────
-
-/// Pipe `output` through the `rtk` CLI subprocess if installed.
-///
-/// Why: When the user has installed RTK (https://github.com/rtk-ai/rtk),
-/// delegating to it gets us the upstream implementation for free, with
-/// updates from the source project. When `rtk` is not on `PATH` we fall
-/// back to the native filter.
-/// What: Spawns `rtk <tool_name>`, writes `output` to stdin, returns stdout.
-/// Returns `None` on any failure (missing binary, non-zero exit, stdin/stdout
-/// IO error, decode error) so the caller can fall back gracefully.
-/// Test: Covered by integration tests when `rtk` is available; unit tests
-/// only verify the `None` path when the binary is absent.
-pub async fn compress_via_rtk(tool_name: &str, output: &str) -> Option<String> {
-    use tokio::io::AsyncWriteExt;
-    use tokio::process::Command;
-
-    // Quick existence check — if `rtk` is not on PATH, skip without spawning.
-    which("rtk")?;
-
-    let mut child = Command::new("rtk")
-        .arg(tool_name)
-        .stdin(std::process::Stdio::piped())
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::null())
-        .spawn()
-        .ok()?;
-
-    if let Some(mut stdin) = child.stdin.take() {
-        // Write output and close stdin so rtk can finish.
-        stdin.write_all(output.as_bytes()).await.ok()?;
-        drop(stdin);
-    }
-
-    let out = child.wait_with_output().await.ok()?;
-    if !out.status.success() {
-        return None;
-    }
-    String::from_utf8(out.stdout).ok()
-}
-
-/// Look up an executable on `PATH`. Returns the absolute path if found.
-///
-/// Why: Avoids depending on the `which` crate while letting us short-circuit
-/// when the binary is absent.
-/// What: Splits `$PATH` (or `;`-separated on Windows), checks `dir/name`
-/// (and `name.exe` on Windows). Returns the first existing match.
-fn which(name: &str) -> Option<std::path::PathBuf> {
-    let path = std::env::var_os("PATH")?;
-    for dir in std::env::split_paths(&path) {
-        let candidate = dir.join(name);
-        if candidate.is_file() {
-            return Some(candidate);
-        }
-        #[cfg(windows)]
-        {
-            let with_ext = dir.join(format!("{name}.exe"));
-            if with_ext.is_file() {
-                return Some(with_ext);
-            }
-        }
-    }
-    None
-}
-
-/// Compress a tool's output, trying the RTK subprocess first and falling
-/// back to the native filter chain.
-///
-/// Why: Most users won't have RTK installed; the native filters are always
-/// available. When RTK is present we delegate so we stay aligned with upstream.
-/// What: Async wrapper — calls `compress_via_rtk`, falls back to
-/// `compress_tool_output` (synchronous, native) on `None`.
-/// Test: `compress_tool_output_async_falls_back_when_rtk_absent`.
-pub async fn compress_tool_output_async(tool_name: &str, output: &str) -> String {
-    if let Some(s) = compress_via_rtk(tool_name, output).await {
-        return s;
-    }
-    compress_tool_output(tool_name, output)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_runner_strips_passing_tests() {
-        let mut input = String::new();
-        for i in 0..10 {
-            input.push_str(&format!("test mod::passing_{i} ... ok\n"));
-        }
-        input.push_str("test mod::failing ... FAILED\n");
-        input.push_str("test result: FAILED. 10 passed; 1 failed\n");
-        let out = filter_test_runner(&input);
-        assert!(out.contains("failing"));
-        assert!(!out.contains("passing_0"));
-        assert!(!out.contains("passing_9"));
-    }
-
-    #[test]
-    fn test_runner_keeps_summary_line() {
-        let input = "test foo ... ok\ntest result: FAILED. 1 passed; 1 failed\n";
-        let out = filter_test_runner(input);
-        assert!(out.contains("test result: FAILED"));
-    }
-
-    #[test]
-    fn test_runner_no_failures_returns_summary_only() {
-        let input = "test a ... ok\ntest b ... ok\ntest result: ok. 2 passed; 0 failed\n";
-        let out = filter_test_runner(input);
-        assert_eq!(out, "test result: ok. 2 passed; 0 failed");
-    }
-
-    #[test]
-    fn test_runner_unknown_tool_passthrough() {
-        let input = "random output line\nanother line\n";
-        let out = compress_tool_output("git_status", input);
-        assert_eq!(out, input);
-    }
-
-    #[test]
-    fn filter_git_diff_strips_context_lines() {
-        let input = "\
---- a/foo.rs
-+++ b/foo.rs
-@@ -1,7 +1,7 @@
- ctx1
- ctx2
- ctx3
--removed
-+added
- ctx4
- ctx5
-";
-        let out = filter_git_diff(input);
-        assert!(!out.contains("ctx1"));
-        assert!(!out.contains("ctx5"));
-        assert!(out.contains("-removed"));
-        assert!(out.contains("+added"));
-        assert!(out.contains("@@ ... @@"));
-    }
-
-    #[test]
-    fn filter_git_diff_preserves_adds_and_removes() {
-        let input = "@@ -1,2 +1,2 @@\n-old line\n+new line\n";
-        let out = filter_git_diff(input);
-        assert!(out.contains("-old line"));
-        assert!(out.contains("+new line"));
-    }
-
-    #[test]
-    fn filter_git_diff_passthrough_no_context() {
-        let input = "--- a/x\n+++ b/x\n@@ -1 +1 @@\n-a\n+b\n";
-        let out = filter_git_diff(input);
-        // No context runs to collapse → unchanged
-        assert_eq!(out, input);
-    }
-
-    #[test]
-    fn filter_git_log_strips_author_date() {
-        let mut input = String::new();
-        for i in 0..10 {
-            input.push_str(&format!("commit abc123{i:03}def456789\n"));
-            input.push_str("Author: Alice <alice@example.com>\n");
-            input.push_str("Date:   Mon Jan 1 12:00:00 2024 +0000\n");
-            input.push('\n');
-            input.push_str(&format!("    feat: subject line {i}\n"));
-            input.push('\n');
-        }
-        // 60 lines total
-        let out = compress_tool_output("git_log", &input);
-        assert!(!out.contains("Author:"));
-        assert!(!out.contains("Date:"));
-        assert!(out.contains("commit abc123"));
-        assert!(out.contains("subject line 0"));
-    }
-
-    #[test]
-    fn filter_git_log_passthrough_short() {
-        // Under 30 lines → passthrough via dispatch
-        let input = "commit abc1234\nAuthor: Bob\nDate: today\n\n    short\n";
-        let out = compress_tool_output("git_log", input);
-        assert_eq!(out, input);
-    }
-
-    #[test]
-    fn filter_file_read_strips_blank_comment_lines() {
-        let mut input = String::new();
-        // 50 code lines + 50 comment lines + 50 blank lines = 150
-        // But we need > 200 for dispatch — generate enough.
-        for i in 0..120 {
-            input.push_str(&format!("let x_{i} = {i};\n"));
-        }
-        for i in 0..60 {
-            input.push_str(&format!("// comment {i}\n"));
-        }
-        for _ in 0..60 {
-            input.push('\n');
-        }
-        let out = compress_tool_output("read_file", &input);
-        assert!(!out.contains("// comment"));
-        assert!(out.contains("let x_0"));
-        assert!(out.contains("let x_119"));
-    }
-
-    #[test]
-    fn filter_file_read_passthrough_short() {
-        let input = "fn main() {\n    println!(\"hi\");\n}\n";
-        let out = compress_tool_output("read_file", input);
-        assert_eq!(out, input);
-    }
-
-    #[test]
-    fn filter_file_read_no_over_filter() {
-        // All-comment file — filtering would leave 0 lines, so return original.
-        let mut input = String::new();
-        for i in 0..250 {
-            input.push_str(&format!("// only comment {i}\n"));
-        }
-        let out = filter_file_read(&input);
-        assert_eq!(out, input);
-    }
-
-    #[test]
-    fn filter_cargo_check_strips_compiling() {
-        let input = "   Compiling foo v0.1.0\n   Compiling bar v0.2.0\n    Finished dev [unoptimized] target(s) in 1.23s\nwarning: unused variable: `x`\n";
-        let out = filter_cargo_check(input);
-        assert!(!out.contains("Compiling"));
-        assert!(!out.contains("Finished"));
-        assert!(out.contains("warning: unused variable"));
-    }
-
-    #[test]
-    fn filter_cargo_check_keeps_warnings() {
-        let input = "   Compiling x v0.1.0\nwarning: foo\nerror: bar\n    Finished\n";
-        let out = filter_cargo_check(input);
-        assert!(out.contains("warning: foo"));
-        assert!(out.contains("error: bar"));
-    }
-
-    #[test]
-    fn compress_tool_output_dispatch_test() {
-        // Inputs must exceed SIZE_GATE_BYTES (80) so the dispatch table runs.
-        // test/cargo (without check/clippy) → test runner filter
-        let test_input = "test alpha ... ok\ntest beta ... ok\ntest gamma ... ok\ntest result: ok. 3 passed; 0 failed\n";
-        let r = compress_tool_output("cargo_test", test_input);
-        assert!(r.contains("test result: ok. 3 passed; 0 failed"));
-        assert!(!r.contains("alpha ... ok"));
-
-        // diff → diff filter
-        let diff_input = "--- a/file.rs\n+++ b/file.rs\n@@ -1,5 +1,5 @@\n ctx1\n ctx2\n-old line of code\n+new line of code\n ctx3\n";
-        let r = compress_tool_output("git_diff", diff_input);
-        assert!(r.contains("@@ ... @@"));
-
-        // unknown → passthrough (small input — size gate, but assertion still holds)
-        let r = compress_tool_output("unknown_tool", "raw output");
-        assert_eq!(r, "raw output");
-
-        // check → cargo_check filter
-        let check_input = "   Compiling foo v0.1.0\n   Compiling bar v0.1.0\n    Finished dev in 1.2s\nwarning: unused variable: `x`\n";
-        let r = compress_tool_output("cargo_check", check_input);
-        assert!(!r.contains("Compiling"));
-        assert!(r.contains("warning"));
-
-        // clippy → cargo_check filter
-        let clippy_input = "   Compiling baz v0.1.0\n    Finished release [optimized] target(s) in 2.34s\nerror: type mismatch in arg\n";
-        let r = compress_tool_output("cargo_clippy", clippy_input);
-        assert!(!r.contains("Finished"));
-        assert!(r.contains("error"));
-    }
-
-    #[test]
-    fn compress_tool_output_reduces_long_passing_test_output() {
-        let mut input = String::new();
-        for i in 0..200 {
-            input.push_str(&format!("test mod::t{i} ... ok\n"));
-        }
-        input.push_str("test result: ok. 200 passed; 0 failed\n");
-        let out = compress_tool_output("cargo_test", &input);
-        assert!(out.lines().count() <= 5);
-    }
-
-    // ── Size gate ────────────────────────────────────────────────────────
-
-    #[test]
-    fn size_gate_skips_short_inputs() {
-        // Input under SIZE_GATE_BYTES is returned unchanged even when the
-        // tool name would otherwise route to a filter.
-        let short = "test foo ... ok\ntest result: ok. 1 passed\n";
-        assert!(short.len() < SIZE_GATE_BYTES);
-        let out = compress_tool_output("cargo_test", short);
-        assert_eq!(out, short, "size gate must passthrough short content");
-    }
-
-    #[test]
-    fn size_gate_lets_large_inputs_through() {
-        // Construct an input over 80 bytes; expect compression to apply.
-        let mut input = String::new();
-        for i in 0..10 {
-            input.push_str(&format!("test passing_{i} ... ok\n"));
-        }
-        input.push_str("test result: ok. 10 passed; 0 failed\n");
-        assert!(input.len() >= SIZE_GATE_BYTES);
-        let out = compress_tool_output("cargo_test", &input);
-        assert!(!out.contains("passing_0 ... ok"));
-    }
-
-    // ── Structured-format detection ──────────────────────────────────────
-
-    #[test]
-    fn is_structured_format_json_object() {
-        assert!(is_structured_format("{\"key\": \"value\", \"n\": 42}"));
-    }
-
-    #[test]
-    fn is_structured_format_json_array() {
-        assert!(is_structured_format("[1, 2, 3, 4]"));
-    }
-
-    #[test]
-    fn is_structured_format_json_with_leading_whitespace() {
-        assert!(is_structured_format("   \n  {\"x\": 1}"));
-    }
-
-    #[test]
-    fn is_structured_format_yaml_doc_marker() {
-        assert!(is_structured_format("---\nname: foo\nversion: 1\n"));
-    }
-
-    #[test]
-    fn is_structured_format_yaml_kv() {
-        assert!(is_structured_format("name: example\nversion: 1.0\n"));
-    }
-
-    #[test]
-    fn is_structured_format_toml_section() {
-        assert!(is_structured_format("[package]\nname = \"foo\"\n"));
-    }
-
-    #[test]
-    fn is_structured_format_csv() {
-        let csv = "id,name,value\n1,foo,10\n2,bar,20\n3,baz,30\n";
-        assert!(is_structured_format(csv));
-    }
-
-    #[test]
-    fn is_structured_format_prose_is_false() {
-        assert!(!is_structured_format(
-            "This is normal prose, not structured data at all."
-        ));
-    }
-
-    #[test]
-    fn is_structured_format_test_output_is_false() {
-        // Test runner output shouldn't be mistaken for structured data.
-        let out = "test mod::foo ... ok\ntest mod::bar ... ok\ntest result: ok\n";
-        assert!(!is_structured_format(out));
-    }
-
-    #[test]
-    fn structured_format_passthrough_via_dispatch() {
-        // A JSON payload routed via a tool name that would otherwise filter
-        // must come back unchanged.
-        let payload = "{\"results\": [{\"name\": \"alpha\", \"status\": \"ok\"}, {\"name\": \"beta\", \"status\": \"fail\"}]}";
-        assert!(payload.len() >= SIZE_GATE_BYTES);
-        let out = compress_tool_output("cargo_test_json", payload);
-        assert_eq!(out, payload);
-    }
-
-    // ── FilterStrategy / Language ────────────────────────────────────────
-
-    #[test]
-    fn language_from_extension_known() {
-        assert_eq!(Language::from_extension("rs"), Language::Rust);
-        assert_eq!(Language::from_extension(".py"), Language::Python);
-        assert_eq!(Language::from_extension("TS"), Language::TypeScript);
-        assert_eq!(Language::from_extension("go"), Language::Go);
-        assert_eq!(Language::from_extension("json"), Language::Data);
-        assert_eq!(Language::from_extension("weird"), Language::Unknown);
-    }
-
-    #[test]
-    fn language_comment_prefix_rust() {
-        assert_eq!(Language::Rust.comment_prefix(), Some("//"));
-        assert_eq!(Language::Python.comment_prefix(), Some("#"));
-        assert_eq!(Language::Data.comment_prefix(), None);
-    }
-
-    #[test]
-    fn language_block_comment_rust() {
-        assert_eq!(Language::Rust.block_comment(), Some(("/*", "*/")));
-        assert_eq!(Language::Python.block_comment(), Some(("\"\"\"", "\"\"\"")));
-        assert_eq!(Language::Data.block_comment(), None);
-    }
-
-    #[test]
-    fn filter_strategy_no_filter_identity() {
-        let f = get_filter(FilterLevel::None);
-        let input = "line one\n\nline two\n// comment\n";
-        assert_eq!(f.filter(input, Language::Rust), input);
-    }
-
-    #[test]
-    fn filter_strategy_minimal_drops_blanks() {
-        let f = get_filter(FilterLevel::Minimal);
-        let input = "line one\n\nline two   \n   \nline three\n";
-        let out = f.filter(input, Language::Unknown);
-        let lines: Vec<&str> = out.lines().collect();
-        assert_eq!(lines, vec!["line one", "line two", "line three"]);
-    }
-
-    #[test]
-    fn filter_strategy_aggressive_strips_rust_line_comments() {
-        let f = get_filter(FilterLevel::Aggressive);
-        let input = "let x = 1;\n// this is a comment\nlet y = 2;\n  // indented comment\n";
-        let out = f.filter(input, Language::Rust);
-        assert!(!out.contains("comment"));
-        assert!(out.contains("let x = 1;"));
-        assert!(out.contains("let y = 2;"));
-    }
-
-    #[test]
-    fn filter_strategy_aggressive_strips_python_hash_comments() {
-        let f = get_filter(FilterLevel::Aggressive);
-        let input = "x = 1\n# hash comment here\ny = 2\n";
-        let out = f.filter(input, Language::Python);
-        assert!(!out.contains("hash comment"));
-        assert!(out.contains("x = 1"));
-        assert!(out.contains("y = 2"));
-    }
-
-    #[test]
-    fn filter_strategy_aggressive_unknown_lang_keeps_comments() {
-        // With no comment prefix known, aggressive becomes minimal.
-        let f = get_filter(FilterLevel::Aggressive);
-        let input = "data1\n// looks like a comment\ndata2\n";
-        let out = f.filter(input, Language::Unknown);
-        assert!(out.contains("// looks like a comment"));
-    }
-
-    // ── RTK subprocess fallback ──────────────────────────────────────────
-
-    #[tokio::test]
-    async fn compress_via_rtk_returns_none_when_binary_absent() {
-        // We can't reliably assert presence in CI, but we CAN assert that
-        // the function returns Some/None without panicking and that the
-        // async fallback always returns a String.
-        let payload = "test result: ok. 100 passed; 0 failed\n".repeat(5);
-        let result = compress_tool_output_async("cargo_test", &payload).await;
-        assert!(!result.is_empty(), "async fallback must return content");
-    }
-
-    #[tokio::test]
-    async fn compress_tool_output_async_falls_back_when_rtk_absent() {
-        // Force-bypass rtk by checking which() with a name guaranteed to not
-        // exist; we test the integration via the public async wrapper.
-        let mut input = String::new();
-        for i in 0..10 {
-            input.push_str(&format!("test t{i} ... ok\n"));
-        }
-        input.push_str("test result: ok. 10 passed; 0 failed\n");
-        let out = compress_tool_output_async("cargo_test", &input).await;
-        // Whether rtk ran or native fallback ran, the summary must be retained.
-        assert!(out.contains("test result"));
-    }
-}
